test[automation]: use counter stringification
